# server/src/router.ts
import express, { Request, Response } from 'express';
import axios from 'axios';
import {
    wastetypes,
    wastpage,
    searchResult,
    recyclingmethod,
} from './types/apiTypes';
import { LeaderboardModel } from './models/Leaderboard';

import dotenv from 'dotenv';
dotenv.config();

const router = express.Router();
const API_BASE_URL = process.env.url;
const headers = {
    client_id: process.env.client_id,
    client_secret: process.env.client_secret,
};

router.get('/wastetypes', async (req: Request, res: Response) => {
    try {
        const { lang } = req.query;
        const response = await axios.get<wastetypes[]>(
            `${API_BASE_URL}/waste-types?lang=${lang}`,
            { headers }
        );
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching waste types:', error);
        throw error;
    }
});

router.get('/wastetypes/:types', async (req: Request, res: Response) => {
    try {
        const { lang } = req.query;
        const { types } = req.params;
        const response = await axios.get<wastpage>(
            `${API_BASE_URL}/waste-types?search=${types}&lang=${lang}`,
            { headers }
        );
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching search results:', error);
        throw error;
    }
});

router.get('/recyclingmethods', async (req: Request, res: Response) => {
    try {
        const { lang } = req.query;
        const response = await axios.get<recyclingmethod[]>(
            `${API_BASE_URL}/recycling-methods?lang=${lang}`,
            { headers }
        );
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching recycling methods:', error);
        throw error;
    }
});

router.get('/recyclingmethods/:method', async (req: Request, res: Response) => {
    try {
        const { lang } = req.query;
        const { method } = req.params;
        const response = await axios.get<recyclingmethod[]>(
            `${API_BASE_URL}/recycling-methods?lang=${lang}&search=${method}`,
            { headers }
        );
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching recycling methods:', error);
        throw error;
    }
});

router.get('/wastepages', async (req: Request, res: Response) => {
    try {
        const { lang } = req.query;
        const page = req.query.page || 1;
        const response = await axios.get<wastpage[]>(
            `${API_BASE_URL}/waste-pages?lang=${lang}&page=${page}`,
            { headers }
        );
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching waste pages:', error);
        throw error;
    }
});

router.get('/search/:q', async (req: Request, res: Response) => {
    try {
        const { q } = req.params;
        const { lang, wasteType, recyclingMethod } = req.query;
        const params = new URLSearchParams({
            q,
            lang: lang as string,
            includeSynonyms: 'true',
        });
        if (wasteType) params.append('wasteType[]', wasteType as string);
        if (recyclingMethod)
            params.append('recyclingMethod[]', recyclingMethod as string);
        const apiUrl = `${API_BASE_URL}/search?${params.toString()}`;
        const response = await axios.get<searchResult[]>(apiUrl, { headers });
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching search results:', error);
        throw error;
    }
});

    router.get(
        '/wastepages/search/:search',
        async (req: Request, res: Response) => {
            console.log('✅ Route hit:', req.params, req.query);
            try {
                const { search } = req.params;
                const { lang, wasteType, recyclingMethod, page } = req.query;
                const params = new URLSearchParams({
                    search,
                    lang: lang as string
                });
                if (page) params.append('page', page as string);
                if (wasteType) params.append('wasteType[]', wasteType as string);
                if (recyclingMethod)
                    params.append('recyclingMethod[]', recyclingMethod as string);
                const apiUrl = `${API_BASE_URL}/waste-pages?${params.toString()}`;
                const response = await axios.get<wastpage[]>(apiUrl, { headers });
                res.json(response.data);
            } catch (error) {
                console.error('Error fetching search results:', error);
                throw error;
            }
        }
    );

router.get('/wastepages/:id', async (req: Request, res: Response) => {
    console.log('single wastepage id route hit');
    try {
        const { lang } = req.query;
        const response = await axios.get<wastpage>(
            `${API_BASE_URL}/waste-pages/${req.params.id}?lang=${lang}`,
            { headers }
        );
        res.json(response.data);
    } catch (error) {
        console.error('Error fetching search results:', error);
        throw error;
    }
});

router.get('/leaderboards', async (req: Request, res: Response) => {
    console.log('leaderboards get route hit');
    try {
        const data = await LeaderboardModel.find().sort({ points: -1 });
        const leaderboards = data.map((item, idx) => ({
            ranking: idx + 1,
            name: item.username,
            points: item.points,
        }));
        res.json(leaderboards);
    } catch (error) {
        console.error('Error fetching leaderboards:', error);
        res.status(500).json({ error: 'Failed to fetch leaderboards' });
    }
});

router.post('/leaderboards', async (req: Request, res: Response) => {
    console.log('leaderboards post route hit');
    try {
        const { username, points } = req.body;

        if (username == null || points == null) {
            return res.status(400).json({ error: 'Username and points are missing' });
        }

        const leaderboardEntry = new LeaderboardModel({
            username,
            points
        })
        await leaderboardEntry.save()

    } catch (error) {
        console.error('Error posting in leaderboards:', error);
        res.status(500).json({ error: 'Failed to post leaderboards' });
    }
});

export default router;